feat: CI build times, low hanging fruits.
This commit does a couple of things that lower CI build times:

  - Separate and run lints / builds / tests in parallel
  - Cache node_modules (in tests), and cargo build output
  - Run cargo test --no-run while we're waiting for the nodejs subservices to be up and running

The overall goal is to let the CI fail as early as possible (lint -> build -> tests).
